refactor: 전역 상태 관심사 별 폴더 구조 분리
