+++ src/components/OrderConfirmPage/CouponItem/CouponItem.tsx
@@ -1,8 +1,8 @@
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { selectedCouponListSelector, isCouponListMaxLength } from '../../../recoil/selectors/selectors';
+import { selectedCouponListSelector, isCouponListMaxLength } from '../../../recoil/Coupon/selectors/selectors';
+import useCouponApplicable from '../../hooks/useCouponApplicable';
 import CheckBox from '../../common/CheckBox/CheckBox';
 import type { Coupon } from '../../../types/Coupon.type';
-import useCouponApplicable from '../../hooks/useCouponApplicable';
 
 import * as S from './CouponItems.style';
