+++ src/pages/OrderConfirmPage/OrderConfirmPage.tsx
@@ -8,8 +8,8 @@ import SubmitButton from '../../components/common/SubmitButton/SubmitButton';
 import CheckBox from '../../components/common/CheckBox/CheckBox';
 import TotalPriceContainer from '../../components/ShoppingCartPage/TotalPriceContainer/TotalPriceContainer';
 import CouponModal from '../../components/OrderConfirmPage/CouponModal/CouponModal';
-import { selectedCartItemListState, isSigolState } from '../../recoil/atoms/atoms';
-import { cartOrderTotalCountSelector } from '../../recoil/selectors/selectors';
+import { selectedCartItemListState, isSigolState } from '../../recoil/CartItem/atoms/atoms';
+import { cartOrderTotalCountSelector } from '../../recoil/CartItem/selectors/selectors';
 import { PATHS } from '../../constants/PATHS';
 import * as S from './OrderConfirmPage.style';
 import { useState } from 'react';
