+++ src/pages/ShoppingCartPage/ShoppingCartPage.tsx
@@ -7,7 +7,7 @@ import SubmitButton from '../../components/common/SubmitButton/SubmitButton';
 import CartItemList from '../../components/ShoppingCartPage/CartItemList/CartItemList';
 import TotalPriceContainer from '../../components/ShoppingCartPage/TotalPriceContainer/TotalPriceContainer';
 import type { TCartItem } from '../../types/CartItem.type';
-import { selectedCartItemListState } from '../../recoil/atoms/atoms';
+import { selectedCartItemListState } from '../../recoil/CartItem/atoms/atoms';
 import { fetchCartItemList } from '../../apis';
 import { EmptyCart } from '../../assets';
 import { PATHS } from '../../constants/PATHS';
