+++ src/recoil/CartItem/atoms/atoms.ts
@@ -1,6 +1,5 @@
 import { atom, AtomEffect } from 'recoil';
-import type { TCartItem } from '../../types/CartItem.type';
-import { Coupon } from '../../types/Coupon.type';
+import type { TCartItem } from '../../../types/CartItem.type';
 
 const localStorageEffect =
   <T>(key: string): AtomEffect<T> =>
@@ -20,18 +19,13 @@ const localStorageEffect =
     });
   };
 
-export const isSigolState = atom<boolean>({
-  key: 'isSigolState',
-  default: false,
-});
-
-export const selectedCouponListState = atom<Coupon[]>({
-  key: 'couponListState',
-  default: [],
-});
-
 export const selectedCartItemListState = atom<TCartItem[]>({
   key: 'selectedCartItemListState',
   default: [],
   effects: [localStorageEffect<TCartItem[]>('selectedCartItemListState')],
+});
+
+export const isSigolState = atom<boolean>({
+  key: 'isSigolState',
+  default: false,
 });
