+++ src/recoil/Coupon/atoms/atoms.ts
@@ -1,0 +1,7 @@
+import { atom } from 'recoil';
+import { Coupon } from '../../../types/Coupon.type';
+
+export const selectedCouponListState = atom<Coupon[]>({
+  key: 'couponListState',
+  default: [],
+});
